wait for builds to complete

+++ src/modules/0.0.22/aws_codebuild/mappers/build.ts
@@ -44,7 +44,7 @@ export class CodebuildBuildListMapper extends MapperBase<CodebuildBuildList> {
     return out;
   }
 
-  async waitForBuildsToStop(client: CodeBuild, ids: string[]) {
+  async waitForBuildsToComplete(client: CodeBuild, ids: string[]) {
     // wait for policies to be attached
     const input: BatchGetBuildsCommandInput = {
       ids,
@@ -61,7 +61,7 @@ export class CodebuildBuildListMapper extends MapperBase<CodebuildBuildList> {
       async (cl, cmd) => {
         try {
           const data = await cl.batchGetBuilds(cmd);
-          const done = data?.builds?.every(bd => bd.buildStatus === BuildStatus.STOPPED || !!bd.endTime);
+          const done = data?.builds?.every(bd => bd.buildStatus !== BuildStatus.IN_PROGRESS || !!bd.endTime);
           if (done) {
             return { state: WaiterState.SUCCESS };
           }
@@ -134,7 +134,7 @@ export class CodebuildBuildListMapper extends MapperBase<CodebuildBuildList> {
         const input: StopBuildInput = { id };
         await this.stopBuild(client.cbClient, input);
       }
-      if (idsToStop.length > 0) await this.waitForBuildsToStop(client.cbClient, idsToStop);
+      if (idsToStop.length > 0) await this.waitForBuildsToComplete(client.cbClient, idsToStop);
       let idsToDel = bds.map(bd => bd.awsId);
       // Wait for ~2.5min until builds can be deleted
       let i = 0;
@@ -178,7 +178,8 @@ export class CodebuildBuildImportMapper extends MapperBase<CodebuildBuildImport>
           projectName: e.project.projectName,
         };
         const cloudBuild = await this.startBuild(client.cbClient, input);
-        if (!cloudBuild) throw new Error('Error starting build');
+        if (!cloudBuild || !cloudBuild.id) throw new Error('Error starting build');
+        await this.module.buildList.waitForBuildsToComplete(client.cbClient, [cloudBuild.id]);
         const dbBuild = await this.module.buildList.buildListMapper(cloudBuild, ctx);
         if (!dbBuild) throw new Error('Error starting build');
         await this.module.buildImport.db.delete(e, ctx);
